requirements.txt
flask
langchain-groq
langchain-community
langchain-core
python-dotenv
requests
langchain